Feat noun optimization - update event names, babylon contract addresses, and state-bank configurations

+++ contracts/delegation-manager/src/msg.rs
@@ -25,7 +25,7 @@ pub struct InstantiateMsg {
     pub withdrawal_delay_blocks: Vec<u64>,
     pub pauser: String,
     pub unpauser: String,
-    pub initial_paused_status: u64,
+    pub initial_paused_status: u8,
 }
 
 #[cw_serde]
