+++ contracts/strategy-base/src/msg.rs
@@ -14,7 +14,7 @@ pub struct InstantiateMsg {
     pub underlying_token: String,
     pub pauser: String,
     pub unpauser: String,
-    pub initial_paused_status: u64,
+    pub initial_paused_status: u8,
 }
 
 #[cw_serde]
