+++ contracts/strategy-factory/src/msg.rs
@@ -7,7 +7,7 @@ pub struct InstantiateMsg {
     pub strategy_manager: String,
     pub pauser: String,
     pub unpauser: String,
-    pub initial_paused_status: u64,
+    pub initial_paused_status: u8,
 }
 
 #[cw_serde]
